UserModel class moved to the userModel file
