+++ src/repository/storage/postgres/orm/sequelize/userList.ts
@@ -1,6 +1,6 @@
 import type User from '@domain/entities/user.js';
 import type { UserList } from '@domain/entities/userList.js';
-import type { UserModel } from '@repository/storage/postgres/orm/sequelize/user.js';
+import type { UserModel } from '@repository/storage/postgres/orm/sequelize/userModel.js';
 import type UserModelSequelizeStorage from '@repository/storage/postgres/orm/sequelize/userModel.js';
 
 /**
